# tracing: Add `spanID` field to access logs and `http.vars.span_id` placeholder