@YifanYang6 2024-10-19
If needed, I can also update the related documentation for these changes in the [Caddy website](https://github.com/caddyserver/website/blob/master/src/docs/markdown/caddyfile/directives/tracing.md).