@mholt 2024-10-19
Thanks for the PR!

Just so I'm clear on this, the performance is only impacted if metrics are enabled, right?