@YifanYang6 2024-10-19
> Thanks for the PR!
> 
> Just so I'm clear on this, the performance is only impacted if metrics are enabled, right?

That's correct. The performance impact occurs only when tracing is enabled, as both `traceid` and `spanid` are added to the logs only in that scenario.