@hairyhenderson 2024-10-21
> Just so I'm clear on this, the performance is only impacted if metrics are enabled, right?

yes if you mean `s/metrics/tracing/` - this is (or, should be) orthogonal to metrics