@digitaltim-de 2025-01-12
No docs for that... I want to use custom tags, any idea how i can send via caddy custom tags to the tracing?

For example i am using this:

```
    route {
        @traceable {
            not path /build/* /images/* /favicon-32x32.png
        }

        handle @traceable {
            tracing {
                span "[{http.request.method}] {http.request.uri.path}"
            }
        }
        
        push
        php_fastcgi php:9000
        encode zstd gzip
        file_server
    }
    ```
    
    I want add tags(); for the tracing any idea how i can do it?