@francislavoie 2025-01-12
All this does is add the span ID to the logs automatically, when tracing is enabled. That's it. There's not really anything to document.

Tagging is a totally different topic, it's not currently a supported feature.